Rework radio button style
