+++ Sources/ProcessOutCoreUI/Sources/DesignSystem/RadioButton/PORadioButtonStyle.swift
@@ -5,10 +5,10 @@
 //  Created by Andrii Vysotskyi on 06.06.2023.
 //
 
-import UIKit
+import SwiftUI
 
 /// Describes radio button style in different states.
-public struct PORadioButtonStyle {
+public struct PORadioButtonStyle: ButtonStyle {
 
     /// Style to use when radio button is in default state ie enabled and not selected.
     public let normal: PORadioButtonStateStyle
@@ -34,6 +34,73 @@ public struct PORadioButtonStyle {
         self.selected = selected
         self.highlighted = highlighted
         self.error = error
+    }
+
+    // MARK: - ButtonStyle
+
+    public func makeBody(configuration: Configuration) -> some View {
+        let style = currentStyle(isPressed: configuration.isPressed)
+        HStack(alignment: .top, spacing: Constants.spacing) {
+            ZStack {
+                Circle()
+                    .fill(Color(style.knob.backgroundColor))
+                Circle()
+                    .strokeBorder(Color(style.knob.border.color), lineWidth: style.knob.border.width)
+                Circle()
+                    .fill(Color(style.knob.innerCircleColor))
+                    .frame(width: style.knob.innerCircleRadius * 2)
+            }
+            .frame(width: Constants.knobSize, height: Constants.knobSize)
+            .offset(
+                y: (knobVerticalOffset(typography: style.value.typography) - Constants.knobSize) / 2
+            )
+            configuration
+                .label
+                .textStyle(style.value)
+                .frame(minHeight: Constants.knobSize)
+        }
+        .padding(.vertical, Constants.minVerticalPadding)
+        .frame(minHeight: Constants.minHeight)
+        .allowsHitTesting(!isSelected)
+    }
+
+    // MARK: - Private Nested Types
+
+    private enum Constants {
+        static let minVerticalPadding: CGFloat = 4
+        static let spacing: CGFloat = 8
+        static let minHeight: CGFloat = 44
+        static let knobSize: CGFloat = 18
+    }
+
+    // MARK: - Private Properties
+
+    @Environment(\.isRadioButtonSelected) private var isSelected
+    @Environment(\.isControlInvalid) private var isInvalid
+
+    @POBackport.ScaledMetric(relativeTo: .body) // todo(andrii-vysotskyi): style is not always body
+    private var contentSizeMultipler: CGFloat = 1
+
+    // MARK: - Private Methods
+
+    private func currentStyle(isPressed: Bool) -> PORadioButtonStateStyle {
+        if isSelected {
+            return selected
+        }
+        if isPressed {
+            return highlighted
+        }
+        if isInvalid {
+            return error
+        }
+        return normal
+    }
+
+    private func knobVerticalOffset(typography: POTypography) -> CGFloat {
+        if typography.textStyle != nil {
+            return typography.lineHeight * contentSizeMultipler
+        }
+        return typography.lineHeight
     }
 }
